mainPresenter unified
